requirements.txt
discord.py
nest_asyncio
aiocron